Ver.2 by Adrian
This code adds the replace book method to the Book Inventory class and includes an option in the main menu to replace a book. When the user chooses to replace a book, they are prompted to enter the index of the book they want to replace, along with the new titles and authors. The method updates the specified book with the new information if the book exists at the given index.